# docs/api.yaml
openapi: 3.0.1
info:
  title: pokerdice API
  description: pokerdice API
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    authCookie:
      type: apiKey
      in: cookie
      name: AuthCookie
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        total: { type: integer, minimum: 0, description: Total de elementos (ignora paginação) }
        limit: { type: integer, minimum: 0 }
        skip: { type: integer, minimum: 0 }
        next: { type: string, format: uri, nullable: true }
        prev: { type: string, format: uri, nullable: true }
        traceId: { type: string, description: ID de correlação para debugging/observabilidade }
    Problem:
      type: object
      required: [type]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }

    # USERS
...
    LobbyPlayer:
      type: object
      properties:
        userId: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        balance: { type: number, format: float }
...
    MatchWithPlayers:
      type: object
      properties:
        matchId: { type: integer }
        lobbyId: { type: integer }
        status: { type: string, enum: [ONGOING, FINISHED] }
        players:
          type: array
          items: { $ref: '#/components/schemas/LobbyPlayer' }
    RoundWithDetails:
      type: object
      properties:
        roundId: { type: integer }
        number: { type: integer }
        blind: { type: number }
        pot: { type: number }
        currentTurn: { $ref: '#/components/schemas/Turn' }
        players:
          type: array
          items: { $ref: '#/components/schemas/LobbyPlayer' }
        hands:
          type: array
          items: { $ref: '#/components/schemas/Hand' }
    Turn:
      type: object
      properties:
        turnId: { type: integer }
        userId: { type: integer }
        number: { type: integer }
        rollCount: { type: integer }
        state: { type: string, enum: [ONGOING, ENDED] }
    Hand:
      type: object
      properties:
        roundId: { type: integer }
        userId: { type: integer }
        diceValues: { type: string, description: "ex: '1,2,3,4,5'" }
        rank: { type: integer }
        tieBreakerKey: { type: integer }
    Roll:
      type: object
      properties:
        rollId: { type: integer }
        turnId: { type: integer }
        createdAt: { type: integer }
        heldMask: { type: string, description: "e.g. '01011'" }
        diceValues: { type: string }
    RollRequestInputModel:
      type: object
      properties:
        heldMask:
          type: string
          description: Obrigatório a partir do 2º lançamento
paths:
...
  /api/matches/{mid}:
    get:
      tags: [Matches]
      summary: Get match
      operationId: getMatch
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/MatchWithPlayers' }
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/matches/{mid}/rounds/current:
    get:
      tags: [Matches]
      summary: Get current round
      operationId: getCurrentRound
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RoundWithDetails' }
        '400':
          description: Invalid match state
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/matches/{mid}/rounds:
    get:
      tags: [Matches]
      summary: List rounds
      operationId: listRounds
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/matches/{mid}/rounds/{rid}:
    get:
      tags: [Matches]
      summary: Get round details
      operationId: getRoundDetails
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
        - in: path
          name: rid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/RoundWithDetails' }
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/matches/{mid}/turns/current:
    get:
      tags: [Matches]
      summary: Get current turn
      operationId: getCurrentTurn
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '400':
          description: Invalid match state
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/matches/{mid}/hands:
    get:
      tags: [Matches]
      summary: Get hand history
      operationId: getHandHistory
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/matches/{mid}/timeline:
    get:
      tags: [Matches]
      summary: Get match timeline (turns)
      operationId: getMatchTimeline
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/matches/{mid}/turns/current/roll:
    post:
      tags: [Matches]
      summary: Roll dice
      operationId: rollDice
      security:
        - bearerAuth: []
        - authCookie: []
      parameters:
        - in: path
          name: mid
          required: true
          schema: { type: integer, format: int32 }
      requestBody:
        required: false   # 1º lançamento não precisa de body
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RollRequestInputModel' }
            examples:
              firstRoll:
                summary: First roll (sem body)
                value: {}
              subsequentRoll:
                summary: Subsequent roll
                value: { heldMask: "11100" }
              invalidHeldMask:
                summary: Subsequent roll (mask inválida / vazia)
                value: { heldMask: "" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Roll' }
        '400':
          description: Invalid state / missing or invalid heldMask
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }